level1 -  changed hardcoded string, moved Register button logic to separate functions, added extension function for toast
