+++ app/src/main/java/com/project/myapp/ExFunctions.kt
@@ -1,0 +1,10 @@
+package com.project.myapp
+
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+
+object ExtensionUtils {
+    fun AppCompatActivity.toast(message: String) {
+        Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+    }
+}
